pyproject.toml
[tool.poetry]
name = "fungi"
version = "0.1.0"
description = "A simple educational P2P project demonstrating the principles of peer-to-peer connectivity and decentralized networking"
authors = ["victor <victorgoubet@orange.fr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pystun3 = "^1.0.0"
pydantic = "^2.7.4"
streamlit = "^1.36.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
launch-fungi = "fungi.launcher:launch"